Rename module rest-util to common-http
