+++ common/common-http/src/main/java/org/ow2/proactive/http/HttpClientBuilder.java
@@ -34,7 +34,7 @@
  * ################################################################
  * $$PROACTIVE_INITIAL_DEV$$
  */
-package org.ow2.proactive.scheduling.util.http;
+package org.ow2.proactive.http;
 
 import java.security.KeyManagementException;
 import java.security.KeyStoreException;
@@ -42,7 +42,7 @@ import java.security.NoSuchAlgorithmException;
 
 import javax.net.ssl.SSLContext;
 
-import org.ow2.proactive.scheduling.util.WebProperties;
+import org.ow2.proactive.web.WebProperties;
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
 import org.apache.http.conn.ssl.SSLContextBuilder;
