+++ rest/rest-cli/src/main/java/org/ow2/proactive_grid_cloud_portal/cli/ApplicationContextImpl.java
@@ -7,7 +7,7 @@ import java.util.Stack;
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
 
-import org.ow2.proactive.scheduling.util.http.HttpClientBuilder;
+import org.ow2.proactive.http.HttpClientBuilder;
 import org.ow2.proactive_grid_cloud_portal.cli.console.AbstractDevice;
 import org.ow2.proactive_grid_cloud_portal.cli.json.PluginView;
 import org.ow2.proactive_grid_cloud_portal.scheduler.client.SchedulerRestClient;
