+++ scheduler/scheduler-server/src/main/java/org/ow2/proactive/utils/JettyStarter.java
@@ -46,7 +46,7 @@ import java.util.Properties;
 import org.objectweb.proactive.core.config.CentralPAPropertyRepository;
 import org.objectweb.proactive.core.util.ProActiveInet;
 import org.ow2.proactive.scheduler.core.properties.PASchedulerProperties;
-import org.ow2.proactive.scheduling.util.WebProperties;
+import org.ow2.proactive.web.WebProperties;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.log4j.BasicConfigurator;
 import org.apache.log4j.Logger;
